Выносит форму авторизации в отдельный компонент.
