+++ project/src/components/pages/sign-in-screen/sign-in-screen.tsx
@@ -1,32 +1,14 @@
-import { ChangeEvent, FormEvent, useState } from 'react';
+
 import { Link } from 'react-router-dom';
 import { AppRoute, CITIES } from '../../../const';
 import { useAppDispatch } from '../../../hooks';
-import { loginAction } from '../../../store/api-actions';
 import { changeCity } from '../../../store/app-data/app-data';
 import { getRandomCity } from '../../../util';
+import Login from '../../login/login';
 
 function SignInScreen ():JSX.Element {
-  const [login, setLogin] = useState<string>('');
-  const [password, setPassword] = useState<string>('');
 
   const dispatch = useAppDispatch();
-
-  const handleSubmit = (evt: FormEvent<HTMLFormElement>) => {
-    evt.preventDefault();
-    if (login && password) {
-      dispatch(loginAction({login, password}));
-    }
-  };
-
-  const loginChange = (evt: ChangeEvent<HTMLInputElement>) => {
-    setLogin(evt.target.value);
-  };
-
-  const passwordChange = (evt: ChangeEvent<HTMLInputElement>) => {
-    evt.target.value = evt.target.value.replace(/\s/g, '');
-    setPassword(evt.target.value);
-  };
 
   const randomCity = getRandomCity(CITIES);
 
@@ -43,42 +25,9 @@ function SignInScreen ():JSX.Element {
           </div>
         </div>
       </header>
-
       <main className="page__main page__main--login">
         <div className="page__login-container container">
-          <section className="login">
-            <h1 className="login__title">Sign in</h1>
-            <form
-              onSubmit={handleSubmit}
-              className="login__form form"
-              action="#"
-              method="post"
-            >
-              <div className="login__input-wrapper form__input-wrapper">
-                <label className="visually-hidden">E-mail</label>
-                <input
-                  onChange={loginChange}
-                  className="login__input form__input"
-                  type="text"
-                  name="email"
-                  placeholder="Email"
-                  required
-                />
-              </div>
-              <div className="login__input-wrapper form__input-wrapper">
-                <label className="visually-hidden">Password</label>
-                <input
-                  onChange={passwordChange}
-                  className="login__input form__input"
-                  type="password"
-                  name="password"
-                  placeholder="Password"
-                  required
-                />
-              </div>
-              <button className="login__submit form__submit button" type="submit">Sign in</button>
-            </form>
-          </section>
+          <Login />
           <section className="locations locations--login locations--current">
             <div className="locations__item">
               <Link
